[Refactor] AdminSearchBar로 대체 #1175
